@@ -14,31 +14,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.1.3",
-    "@angular/common": "~10.1.3",
-    "@angular/compiler": "~10.1.3",
-    "@angular/core": "~10.1.3",
+    "@angular/animations": "~10.2.1",
+    "@angular/common": "~10.2.1",
+    "@angular/compiler": "~10.2.1",
+    "@angular/core": "~10.2.1",
     "@angular/fire": "../dist/packages-dist",
-    "@angular/forms": "~10.1.3",
-    "@angular/platform-browser": "~10.1.3",
-    "@angular/platform-browser-dynamic": "~10.1.3",
-    "@angular/platform-server": "~10.1.3",
-    "@angular/router": "~10.1.3",
-    "@angular/service-worker": "^10.1.3",
+    "@angular/forms": "~10.2.1",
+    "@angular/platform-browser": "~10.2.1",
+    "@angular/platform-browser-dynamic": "~10.2.1",
+    "@angular/platform-server": "~10.2.1",
+    "@angular/router": "~10.2.1",
+    "@angular/service-worker": "^10.2.1",
     "@nguniversal/express-engine": "~10.1.0",
-    "firebase": "^7.21.1",
+    "firebase": "^8.0.0",
     "first-input-delay": "^0.1.3",
     "proxy-polyfill": "^0.3.2",
     "rxjs": "~6.6.3",
     "tslib": "^2.0.1",
-    "zone.js": "0.10.2"
+    "zone.js": "0.10.3"
   },
   "devDependencies": {
     "@angular-devkit/architect": "~0.1001.3",
-    "@angular-devkit/build-angular": "~0.1001.3",
-    "@angular/cli": "~10.1.3",
-    "@angular/compiler-cli": "~10.1.3",
-    "@angular/language-service": "~10.1.3",
+    "@angular-devkit/build-angular": "~0.1002.0",
+    "@angular/cli": "~10.2.0",
+    "@angular/compiler-cli": "~10.2.1",
+    "@angular/language-service": "~10.2.1",
     "@firebase/app-types": "^0.6.1",
     "@nguniversal/builders": "^10.1.0",
     "@types/jasmine": "~3.5.0",
