@@ -1,6 +1,6 @@
 import { Injectable, Inject, Optional, NgZone, PLATFORM_ID } from '@angular/core';
 import { Observable, of, from } from 'rxjs';
-import { switchMap, map, observeOn, shareReplay, first, tap } from 'rxjs/operators';
+import { switchMap, map, observeOn, shareReplay, first } from 'rxjs/operators';
 import {
   FIREBASE_OPTIONS,
   FIREBASE_APP_NAME,
@@ -12,10 +12,10 @@ import {
   ɵAngularFireSchedulers,
   ɵkeepUnstableUntilFirstFactory
 } from '@angular/fire';
-import { User, auth } from 'firebase/app';
+import firebase from 'firebase/app';
 import { isPlatformServer } from '@angular/common';
 
-export interface AngularFireAuth extends ɵPromiseProxy<auth.Auth> {}
+export interface AngularFireAuth extends ɵPromiseProxy<firebase.auth.Auth> {}
 
 @Injectable({
   providedIn: 'any'
@@ -25,7 +25,7 @@ export class AngularFireAuth {
   /**
    * Observable of authentication state; as of Firebase 4.0 this is only triggered via sign-in/out
    */
-  public readonly authState: Observable<User|null>;
+  public readonly authState: Observable<firebase.User|null>;
 
   /**
    * Observable of the currently signed-in user's JWT token used to identify the user to a Firebase service (or null).
@@ -35,14 +35,14 @@ export class AngularFireAuth {
   /**
    * Observable of the currently signed-in user (or null).
    */
-  public readonly user: Observable<User|null>;
+  public readonly user: Observable<firebase.User|null>;
 
   /**
    * Observable of the currently signed-in user's IdTokenResult object which contains the ID token JWT string and other
    * helper properties for getting different data associated with the token as well as all the decoded payload claims
    * (or null).
    */
-  public readonly idTokenResult: Observable<auth.IdTokenResult|null>;
+  public readonly idTokenResult: Observable<firebase.auth.IdTokenResult|null>;
 
   constructor(
     @Inject(FIREBASE_OPTIONS) options: FirebaseOptions,
@@ -76,14 +76,14 @@ export class AngularFireAuth {
       const _ = auth.pipe(first()).subscribe();
 
       this.authState = auth.pipe(
-        switchMap(auth => auth.getRedirectResult().then(() => auth)),
-        switchMap(auth => zone.runOutsideAngular(() => new Observable<User|null>(auth.onAuthStateChanged.bind(auth)))),
+        switchMap(auth => auth.getRedirectResult().then(() => auth, () => auth)),
+        switchMap(auth => zone.runOutsideAngular(() => new Observable<firebase.User|null>(auth.onAuthStateChanged.bind(auth)))),
         keepUnstableUntilFirst
       );
 
       this.user = auth.pipe(
-        switchMap(auth => auth.getRedirectResult().then(() => auth)),
-        switchMap(auth => zone.runOutsideAngular(() => new Observable<User|null>(auth.onIdTokenChanged.bind(auth)))),
+        switchMap(auth => auth.getRedirectResult().then(() => auth, () => auth)),
+        switchMap(auth => zone.runOutsideAngular(() => new Observable<firebase.User|null>(auth.onIdTokenChanged.bind(auth)))),
         keepUnstableUntilFirst
       );
 
