@@ -12,7 +12,6 @@ import {
   ɵfirebaseAppFactory,
   ɵPromiseProxy
 } from '@angular/fire';
-import { analytics } from 'firebase/app';
 import firebase from 'firebase/app';
 
 export interface Config {
@@ -33,12 +32,12 @@ const GTAG_CONFIG_COMMAND = 'config';
 const GTAG_FUNCTION_NAME = 'gtag';
 const DATA_LAYER_NAME = 'dataLayer';
 
-export interface AngularFireAnalytics extends ɵPromiseProxy<analytics.Analytics> {
+export interface AngularFireAnalytics extends ɵPromiseProxy<firebase.analytics.Analytics> {
 }
 
 let gtag: (...args: any[]) => void;
 let analyticsInitialized: Promise<void>;
-const analyticsInstanceCache: { [key: string]: Observable<analytics.Analytics> } = {};
+const analyticsInstanceCache: { [key: string]: Observable<firebase.analytics.Analytics> } = {};
 
 @Injectable({
   providedIn: 'any'
@@ -66,6 +65,12 @@ export class AngularFireAnalytics {
     if (!analyticsInitialized) {
       if (isPlatformBrowser(platformId)) {
         window[DATA_LAYER_NAME] = window[DATA_LAYER_NAME] || [];
+        /**
+         * According to the gtag documentation, this function that defines a custom data layer cannot be
+         * an arrow function because 'arguments' is not an array. It is actually an object that behaves
+         * like an array and contains more information then just indexes. Transforming this into arrow function
+         * caused issue #2505 where analytics no longer sent any data.
+         */
         // tslint:disable-next-line: only-arrow-functions
         gtag = (window[GTAG_FUNCTION_NAME] as any) || (function(..._args: any[]) {
           (window[DATA_LAYER_NAME] as any).push(arguments);
