@@ -11,7 +11,6 @@ import {
   ɵkeepUnstableUntilFirstFactory
 } from '@angular/fire';
 import { UploadMetadata } from './interfaces';
-import { storage } from 'firebase/app';
 import 'firebase/storage';
 import firebase from 'firebase/app';
 import { registerStorage } from '@firebase/storage';
@@ -29,7 +28,7 @@ export const BUCKET = new InjectionToken<string>('angularfire2.storageBucket');
   providedIn: 'any'
 })
 export class AngularFireStorage {
-  public readonly storage: storage.Storage;
+  public readonly storage: firebase.storage.Storage;
 
   public readonly keepUnstableUntilFirst: <T>(obs: Observable<T>) => Observable<T>;
   public readonly schedulers: ɵAngularFireSchedulers;
